make: Ensure DTB dir is prebuilt and contains DTBs to avoid empty dtb.img
We should be making sure we found .dtb files, because if the directory
doesn't exist for some reason, or there aren't any DTBs in it, an empty
dtb.img will be generated. Warn the user if there are no DTBs found.

Also ensure prebuilt DTB directory does not point to a directory in out/.
The prebuilt directory to dtb.img logic does not accomodate for dtbs that
are not present before the build is started, thus the generated image will
most likely be wrong. Error if the prebuilt dir path is in the out dir.

Test: Remove device/google/bonito-kernel and attempt to build;
      Check build warning due to empty DTB list
Test: Remove device/google/bonito-kernel/sdm670.dtb and build;
      Check build warning due to empty DTB list
Test: Set BOARD_PREBUILT_DTBIMAGE_DIR to
      out/target/product/sargo/obj/KERNEL_OBJ/arch/arm64/boot/dts/qcom
      Check build fails due to BOARD_PREBUILT_DTBIMAGE_DIR being in out/
Test: Build b4s4 and b1c1 dtb images before and after change and
      verify the same image is produced.
Change-Id: Ia9336e9767999bc792aa977a73ae9739a4bc2c37