releasetools: Support repeatable builds.
This CL updates the computation of ext4 FS uuid and AVB salt, to be
based on BuildInfo.fingerprint. This allows getting identical images
when there's no change to image files. We used to support the same prior
to Q, but we unintentionally broke it when moving away from build-time
fingerprint.

Bug: 134525174
Test: TreeHugger
Test: lunch a target; `atest --host releasetools_test releasetools_py3_test`
Test: `m dist`; `unzip out/dist/target_files.zip IMAGS/\*`;
      `zip -d out/dist/target_files.zip IMAGS/\*`
Test: `add_img_to_target_files out/dist/target_files.zip`. Verify that
      the rebuilt images are identical to the ones extracted above.
Change-Id: I8844fc7f044e9abfcd96a00e8958fa0583c70eaf