Update patch set 1
Patch Set 1:

If the build.prop contains multiple lines which have the same key but with the different values, such as "ro.build.fingerprint=abc" and "ro.build.fingerprint=xyz", the "LoadDictionaryFromLines" save the value with "xyz" for the key "ro.build.fingerprint".

But the system in the phone get the value is "abc", that is the problem.

Patch-set: 1