Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > What's the motivation for keeping zip comments in an APK?
> > 
> > Some other third party apks may be integrated into the system as the prebuilt apks. And the zip comment information usually represents the distribution channel information for the app.
> > As you known, we usually don't modify information about these prebuilts apps.
> 
> We are inherently modifying the package if we are calling zipalign on it.
> 
> If you want to keep the comment I'd suggest doing it inside zipalign.

It sounds good. But I still can't fully understand the logic of 'zipalign'. So I don’t dare to rush. I considered that this patch is the least risky and minimal change. I am modifying it in the 'device/xx/build/core/definitions.mk' for specific maintenance in a specific project and it is very easy to transplant.

Patch-set: 1