Update patch set 3
Patch Set 3:

I'd suggest taking this CL out of the topic and submitting the other one on its own (with the Merged-In tag). This comment/cleanspec stuff is not necessary to revert (the comment actually applies with the other CL reverted anyway).

Patch-set: 3
Reviewer: Gerrit User 1060959 <1060959@85c56323-6fa9-3386-8a01-6480fb634889>