Update patch set 4
Patch Set 4:

(3 comments)

> Patch Set 4:
> 
> @elsk,
> BTW, is this intended that sign_target_files_apks does not build signed super.img again for devices launching with Android 10 & A/B compatible? What we do for generating release images is 'm dist -> sign_target_files_apks(LG Key signing) -> add_img_to_target_files(for some additional process after signing)'. Currently, signed super.img seems to be only generated for retrofit devices. Am I on the wrong track?

Signed super.img seems to be only generated for retrofit devices because it is needed to generate the OTA package. You should use individual images whenever possible.

For launch devices, the super image is generated only for development purposes, so it is intentional that it is not signed.

Patch-set: 4
Reviewer: Gerrit User 1121838 <1121838@85c56323-6fa9-3386-8a01-6480fb634889>