Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+1
> 
> (1 comment)
> 
> > (The same as before, this is just adding to our tech debt) If these shouldn't be installed, then the build/installation shouldn't exist. We're still wasting time generating the build rules and building these in checkbuilds.
> 
> I think everyone agrees that this is a short/mid-term kludge which should go away with b/128708192 and b/129006418. However, even then I think the find check is worthwhile, although it ought to be generated from Soong.

I also agree that this should not be implemented this way and we should rely on some Soong logic to prevent these files from being built/installed. However the support is not there yet, and the ETA is not clear to me. Meanwhile, we want to catch regressions like b/142944799 (an issue which was discovered by accident), even if this requires temporary hacks like these.

Patch-set: 4
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0