Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > It looks like we've now got all the support (in apex_available) to actually remove the build/install rules (right Jiyong?). Could we mark all of these modules w/apex_available (just the necessary apex[es], not including //apex_available:platform) first so that we only hit these checks if something really really goes wrong?
> 
> I tried to quickly try this out, and it looks like apex_available may not play well with hostdex or using the java libraries statically (which causes `m checkbuild` to quickly fail with the dangling dependencies check).
> 
> (And really +Jiyong this time)

Right. It right now doesn't handle static dependency, but it's a fixable problem. (btw, failing to track the static dependency doesn't seem to be a big problem for you as this check here also doesn't handle the case). Having this check here in main.mk makes me feel uncomfortable, especially when I see ART library names in the file. ART is a special module, but I think it shouldn't be treated as special at least in this context.

Patch-set: 7