Use NetworkStackNext as preinstalled network stack
NetworkStackNext is the build of the network stack targeted at the
next API, e.g. system_current and not system_29.

While NetworkStack is a mainline module that needs to be built with a
stable SDK (system_29), NetworkStackNext is necessary to test
functionalities that require the in-development SDK (system_current).
Using NetworkStackNext preinstalled on development builds allows for
testing the new APIs and functionalities.

This causes all builds to use com.android.networkstack, which is not a
mainline module package (only the NetworkStack target has its package
renamed, not NetworkStackNext). This must only apply to development
branches.

Once system_current is finalized (system_30 ?), NetworkStack can start
using it, and the system image should switch back to NetworkStack
instead of NetworkStackNext.

Bug: 139269711
Test: Built, phone boots, WiFi working
Change-Id: I885b2c441b99daa3165a4a365738682958478291