Add /persist symlink to BoardConfigMainlineCommon.mk
Since /persist is a SoC specfic symlink, it must not be included in
the root directory. For this reason, we already moved the directory
under /mnt/vendor. However, there are still many modules that are
using the old path /persist.
Until we clear all these violations, we need to have the symlink in
the root directory.

Bug: 143732851
Test: build and check boot and basic functions
Change-Id: Iaee28ba29f79f1c286e090f97173e3196d2fc823