Add TARGET_FLATTEN_APEX_TO_SYSTEM_EXT
if the TARGET_FLATTEN_APEX_TO_SYSTEM_EXT is true and TARGET_FLATTEN_APEX
is false, the flattened apex (if it will be installed on system
partition)  is installed on system_ext partition. But if the
TARGET_FLATTEN_APEX is true, all flattened apexes are installed on the
configured partition.

Test: 1. Add TARGET_FLATTEN_APEX := true and
         PRODUCT_PACKAGES += com.android.art.debug.flattened.
	 Check com.android.art.debug.flattened apex on system_ext
	 partition.
      2. m -j
Bug: 139053989

Change-Id: I4f121be7e42309e1cf87df2a13d7a05f32ad7e8d