Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (2 comments)
> 
> > Patch Set 1: Code-Review+2
> > 
> > This is a GSI-only flag. Is there anything we can do to restrict this to GSI?
> 
> I think we can't do it because I didn't find a special flag in https://android.googlesource.com/platform/build/+/refs/heads/master/target/product/gsi_arm64.mk
> 
> But I'll contact to GSI owner for it.

I know this is not elegant, but can we check it by the name of the product or device? for example, if this is turned on for the device name that doesn't start with gsi_, then we could trigger an error.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>