Improve checks for include dirs outside the source tree
Include directories are expected to be within the source tree, otherwise
it's possible for things that are versioned separately (outside the
source tree) to affect the source tree.

In general, that's not even been the intent for the cases that I've
found that hit this check -- they're not trying to use headers outside
the source tree. They're just bad makefiles that are wrongly assuming
that LOCAL_C_INCLUDES takes relative paths from the current directory.
In that case, if the builds were working before, it's safe to just
remove the offending entry, as it wasn't actually finding the headers
anyways. This is already an error in Android.bp files.

The other case that we already were checking for is more harmful --
using absolute paths, even into the same source tree, makes your builds
unreproducible, as the path to the source tree will now be embedded
within the binaries.

So as a temporary relief, I've added a BUILD_BROKEN_OUTSIDE_INCLUDE_DIRS
that can be set to `true` temporarily to turn this error into a warning.
I don't expect this warning to last more than a single release -- it's
fairly simple to clean up these bad makefiles.

Test: treehugger (build_test on downstream branches)
Change-Id: I6ad6b454a6fccf93772fda26cfc7dd6bbfc07c40