apex-libs-absence checks are removed
The check is implemented in Soong via the apex_availability property.
For a module that should be in the APEX named foo and shouldn't be in
any other APEX and also in the platform (the non-uptadable part), the
property can be set to "foo" to express the restriction and then Soong
will enforce it.

Bug: 128708192
Test: m
Change-Id: Ia1aaaacd685f466447b61deae2849cb0aa83def3