Add PRODUCT_IS_GSI flag for GSI targets
This may not be a good idea but some GSI-specific build rules can be
exploit this. For example, GSI targets are supposed to have both
'flattened' and 'unflattened' APEXes. Without this variable, we have to
keep a list for APEXes separately from PRODUCT_PACKAGES, which would
make error-prone when adding a new APEX.

Additionally, turned off "TARGET_FLATTEN_APEX" explicitly for GSI
because otherwise it will be 'true' by default and then only "flattened"
APEXes will be installed.

Bug: 137802149
Test: lunch aosp_arm64-userdebug && m
  resulting apex images under /system/apex
  and flattened apexes under /system/system_ext/apex

Change-Id: Ib72d1b9dbd59727942da39323ee7e60ac6e14882