Update patch set 4
Patch Set 4:

> Patch Set 2: Code-Review+1
> 
> What's the state of this?

Hi Anton,

This topic was de-prioritized because of the Jan IC. I just restarted to work on it. Now I am waiting for the API review.

Since there are merge conflicts with git_master, after the API review in AOSP, I will first upload and check in these changes in git_master and then AOSP.

Thanks

Patch-set: 4