Pass signing_args to apexer to repack apex
In the build signing environment, we don't necessary have access to the
physical private key. And we work around it by passing an extra signing_helper
to the avbtool instead of a private key. Since we need to perform signing
to add the hashtree footer when rebuilding the apex image. We should propogate
the signing_helper to apexer and eventually to avbtool.

Bug: 148627666
Test: unit tests pass. Smoke test by pass a public key to apexer and
hardcode the private key in signing helper script.
Change-Id: I56d57485493a5ba2cbe4c49e493ee9604f68088b