Update patch set 2
Patch Set 2:

It should be relatively unusual for libraries to be in this list, they're usually included based on what the binaries/etc need. The exceptions are for dynamic dlopens, or APIs that expect to be provided to other partitions / execution environments (apps; vendors; etc). Does this fall under one of those categories?

Patch-set: 2
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>