Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > can you give an example of the ranges...
> > 
> > Here's an example from product.map with e2fsdroid fixed to support holes:
> > 
> > /product/app/Chrome/Chrome.apk 8186-8446 0 8447-11393
> > 
> > Meaning logical blocks 1-260 map to physical blocks 8186-8446, and logical block 261 maps to nothing.
> > 
> > When holes aren't included in the base map, it's much more difficult to preserve the correct logical to physical mapping. See aosp/1236127 for examples.
> 
> 
> It's intuitively weird. I believe when we construct the ranges() class, the ranges has been sorted to 0 8186-8446 8447-11393. And if you subtract the range(0), we are in fact subtracting the 1st block?
> 1. Is there any cases where block 0 is needed in the sparse image? 
> 2. Should we instead amend the holes to make the range 8186-11393....

The ranges are read from block map files, which are intended to preserve logical to physical block mappings. So, recording holes there makes sense. It's the same information that's found in the extent tree, which is what BaseFS is supposed preserve across builds.

Amending the hole would mean e2fsdroid has to allocate physical blocks for those positions, which is wasting space if no block is needed. It's not a lot of space, by a rough estimate there are about ~3MB of holes in Pixel's product partition. But no reason to waste it if we don't have to.

It does look like RangeSet throws away the logical ordering and sorts the block ranges. Does that matter for set operations, though? I'm just trying to strip holes from the list, since "clobbered_blocks" always includes block 0.

I could also use a different, non-numeric marker for holes, but that'll be a significantly bigger change (and RangeSet would have to filter the markers out).

Patch-set: 2