Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: -Code-Review
> > 
> > -code review for clarification.  Oliver, have you tested this on the test machines?  The tool also depends on libc++.so and will fail to run if the test machines don't have libc++ installed.
> 
> No, this doesn't appear to work without libc++.so.1.  Is there a way to build this from source without any external dependencies, or should this library also be added as a build artifact?

One option is to copy llvm-profdata and libc++.so (1.3M each) into build artifacts.  Could be an easy/robust solution to put in place.

We should rethink this as we deploy coverage more broadly (for instance, during presubmit).
Option 1: Since a new llvm-profdata can read older profiles, we can keep an always up-to-date llvm-profdata in a path accessible to the coverage pipeline.

Option 2: We could do as option 1 but keep one for each clang version.  The version used for a build can be pulled from the build logs and used to find the right llvm-profdata.  This is more robust than option 1 since we won't accidentally introduce errors/regressions.

Patch-set: 3
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0