Trim down dont_bother_goals
This had caused the nodeps targets to break when I added the
auto-cleanup feature, as we wrote out smaller lists of what was
installed. The functionality of dont_bother_goals is also being fixed in
https://android-review.googlesource.com/c/platform/build/+/1250929/

But there's also no reaons to have all of these targets in
dont_bother_goals, as they can just re-use the cached kati build graph.

Bug: 149460609
Test: m systemimage && m snod    <did not delete everything>
Test: m snod    <still see warning>
Test: codesearch for uses of MAKECMDGOALS and these targets
Change-Id: Ib8e7ba67e5ad980be4bdcf521c5156eea7941f97