Fix the package name during apex repacking
When doing apk-in-apex signing, the package name is not reserved
during the apex repacking. As a result, the name accidentally
reverts to 'com.android.wifi' from 'com.google.android.wifi'.

This cl extracts the package name from 'aapt' and passed to apexer
with package name override. We expect to find a better way to handle
the binary xml AndroidManifest.xml later.

Test: 152084536
Bug: unit tests pass
Change-Id: I8332b2ee84832fb196f2e1c4309abac5ab92e153