strings.mk: Allow empty RHS values for inputs of collapse-pairs
Currently, when we pass a list of the form "a=b c= d=e" to the function
collapse-pairs, it returns the collapsed list "a=b c=d=e".
As the function is used (only) for formatting property files, this
behavior may result in an unexpectedly ill-formed property file which
is hard to debug (e.g. in the above example, the property "c", which
originally has no value, is interpreted as having a nonempty value "d=e"
whereas the property "d", which originally has the value "e", is missing
from the resulting file).
This CL revises the function so as to return empty properties as is
(e.g. it returns "a=b c= d=e" for the above input), while preservng the
intended behavior for well-formed lists like "a=b c= d e = f".

Bug: 152379493
Test: make
Change-Id: I35faeaedc3bc42e56e01201baf7ea6805a610439