Update patch set 1
Patch Set 1:

Hi Colin, this change creates a new optional output txt that lists every time that a makefile in dir1 tries to find-files-in-subdirs in dir2, when dir1 != dir2.

The purpose of this file is to expose this dependency from dir1 to dir2 when it isn't otherwise exposed by a build rule dependency.

If dir2 doesn't exist in a partner's source tree, then a makefile in dir1 that tries to `$(find-files-in-subdirs, dir2, ...)` will fail when it tries to `cd` to dir2. This .txt exposes that dependency so that partner tooling knows that dir2 needs to be included in the source tree if dir1 is included.

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>