Postpones find-files-in-subdirs errors when ALLOW_MISSING_DEPS.
When ALLOW_MISSING_DEPENDENCIES=true, errors from a nonexistent base
dir now appear as a Ninja error, and only appear if the module using
the result of this function call is built.

A warning is printed for any missing directory, even if the module is
not built.

This allows users to ignore missing base dirs for modules that they do
not build.

Test: (as described below)
  1. Replace an existing find-files-in-subdirs call
     with a nonsense base dir.
  2. m <module that uses those files>
     (UNCHANGED) Observe the standard
     'cd: <nonsense path>: No such file or directory'
     before the Ninja phase begins.
  3. ALLOW_MISSING_DEPENDENCIES=true m <module>
     (NEW) Observe warning, then the Ninja-phase build error
     'ninja: path/to/base_dir_does_not_exist missing
      and no known rule to make it'
  4. ALLOW_MISSING_DEPENDENCIES=true m <unrelated module>
     (NEW) Observe warning but no build error.

Bug: 151670189
Change-Id: Id47d542e2c7d418a22710c909d8e7b72081ef15f