Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > These are intended to be errors in order to catch issues at the time of the problem instead of debugging arbitrary issues later. If we were to consider this, I'd likely prefer optionally removing `--werror_find_emulator` (but that's a regression) rather than adding this overhead to every user.
> 
> This is specifically for partner trees that are attempting to do partial manifests, effectively ALLOW_MISSING_DEPENDENCIES for find-files-in-subdirs.  Maybe it should reuse ALLOW_MISSING_DEPENDENCIES instead of adding a new flag?
> 
> If you rewrote it with an ifdef outside the define there would be no overhead:
> 
>   ifndef ALLOW_MISSING_DEPENDENCIES
>     define find-files-in-subdirs
>       <original implementation>
>     endef
>   else
>     define find-files-in-subdirs
>       <lazy implementation>
>     endef
>   endif

But ALLOW_MISSING_DEPENDENCIES won't cause you to lose half of your source files and continue trying to build the module, while this could (as the module isn't disabled in make, unlike in Soong where we could do that, but we largely fix this problem by requiring non-local sources to be module references).

The multiple implementations would be an improvement, but it should still warn when it can't find the source files, since that's your only indication that anything could be wrong.

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>