Update patch set 1
Patch Set 1:

This is still blocked on https://r.android.com/1292863 (former http://ag/11212475 was considered infeasible because of the required modifications in partner's code). I'm working on it, but it may take a few more days as I need to rework boot image paths generation in Soong. Sorry for the delay.

Patch-set: 1
Reviewer: Gerrit User 1550939 <1550939@85c56323-6fa9-3386-8a01-6480fb634889>