Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The cl doesn't handle the common case where the vendor image isn't a prebuilt image.... In general, we want to update a few files under vendor/ and rebuild vendor.img.
> 
> So you can add some indication in the metadata saying the vendor is a prebuilt. And make the fix in your own branch.

Hi Tianjie,
thank you for reviewing this change. I wanted to ask for some clarification. Can you explain what metadata file you're saying I should be adding indication that the vendor.img is prebuilt?
Also, if there is a vendor.img under IMAGES is it possible that there could be a VENDOR folder too? Is this the case you are saying is not handled by this patch?

Patch-set: 1