Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (2 comments)
> 
> > Patch Set 2:
> > 
> > I started getting rid of this as well:
> > 
> > $(call inherit-product, $(SRC_TARGET_DIR)/product/emulator_vendor.mk)
> > 
> > but that causes issues in the mainline module build which would require touching the makefile in there to fix it, so I'd rather get this in first as it is breaking the GKI.
> 
> What is the problem? It shouldn't cause issues for the mainline module build. If it's complaining about artifact path requirements, those can be easily fixed.

Yes it is an artifact path requirement. I didn't know offhand how to solve that. Even if it is an easy fix, I'd really like to just get this in ASAP to fix the GKI, which is broken on ci.android.com right now.

Patch-set: 2
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0